Feat: 몸통 출력 기능 구현
빼빼로의 몸통 부분을 출력할 수 있도록 기능을 추가함.
