+++ CodeStarterCamp_Week1/main.swift
@@ -8,27 +8,19 @@
 
 import Foundation
 
-func drawIcecream(iceWidth: Int, iceHeight: Int) {
-    for _ in 1...iceHeight {
-        for _ in 1...iceWidth {
-            print("*", terminator: "")
-        }
-        print()
-    }
-    drawStick(topWidth: iceWidth)
-}
+var bodyLength: Int = 12
+var bodyShape: String = "|0|"
+var topping: String = ""
+var stickLength: Int = 4
 
-func drawStick(topWidth: Int) {
-    for _ in 1...4 {
-        for _ in 1...(topWidth/2 - 1) {
-            print(" ",terminator: "")
-        }
-        if topWidth % 2 == 1 {
-            print("| |")
+func drawBody(count: Int) {
+    if topping.isEmpty {
+        print(" \(bodyShape)")
+    } else {
+        if count % 2 == 1 {
+            print("\(topping)\(bodyShape)")
         } else {
-            print("||")
+            print(" \(bodyShape)\(topping)")
         }
     }
 }
-
-drawIcecream(iceWidth: 11,iceHeight: 8)
